third_party/WebKit/LayoutTests/platform/win/external/wpt/XMLHttpRequest/send-accept-language-expected.txt

This is a testharness.js-based test.
FAIL Send "sensible" default value, whatever that means assert_regexp_match: expected object "/Accept-Language:\s.+/" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
FAIL XMLHttpRequest: send() - Accept-Language assert_equals: expected "Accept-Language: x-GameSpeak\n" but got "{\"error\": {\"message\": \"Traceback (most recent call last):\\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\Tools\\Scripts\\webkitpy\\thirdparty\\wpt\\wpt\\tools\\wptserve\\wptserve\\handlers.py\\", line 246, in __call__\n    rv = self.func(request, response)\n  File \\"E:\\b\\c\\b\\win_layout\\src\\third_party\\WebKit\\LayoutTests\\external\\wpt\\XMLHttpRequest/resources/inspect-headers.py\\", line 31, in main\n    result = get_response(request.raw_headers, filter_value, filter_name)\nAttributeError: 'Request' object has no attribute 'raw_headers'\n\", \"code\": 500}}"
Harness: the test ran to completion.
